Extract users db logic to db layer
